// src/client/PacketID.ts
export const enum PacketID {
  // Handshaking C→S
  HANDSHAKING_INTENTION = 0x00,

  // Login S→C
  SERVER_LOGIN_DISCONNECT = 0x00,
  SERVER_HELLO = 0x01,
  SERVER_LOGIN_FINISHED = 0x02,
  SERVER_LOGIN_COMPRESSION = 0x03,

  // Login C→S
  CLIENT_HELLO = 0x00,
  CLIENT_KEY = 0x01,
  CLIENT_LOGIN_ACKNOWLEDGED = 0x03,

  // Configuration S→C
  SERVER_DISCONNECT = 0x02,
  SERVER_FINISH_CONFIGURATION = 0x03,
  SERVER_KEEP_ALIVE = 0x04,
  SERVER_RESOURCE_PACK_PUSH = 0x09,
  SERVER_SELECT_KNOWN_PACKS = 0x0E,

  // Configuration C→S
  CLIENT_FINISH_CONFIGURATION = 0x03,
  CLIENT_KEEP_ALIVE = 0x04,
  CLIENT_RESOURCE_PACK = 0x06,
  CLIENT_SELECT_KNOWN_PACKS = 0x07,

  // Play S→C
  SERVER_DISCONNECT_PLAY = 0x20,
  SERVER_KEEP_ALIVE_PLAY = 0x2B,
  SERVER_LOGIN = 0x30,
  SERVER_RESOURCE_PACK_PUSH_PLAY = 0x4F,
  SERVER_START_CONFIGURATION = 0x74,
  SERVER_TRANSFER = 0x7F,

  // Play C→S
  CLIENT_CONFIGURATION_ACKNOWLEDGED = 0x0F,
  CLIENT_KEEP_ALIVE_PLAY = 0x1B,
  CLIENT_RESOURCE_PACK_PLAY = 0x30,
}